reintroduce Nix.Utils, pass ith through Prelude
After making this flexible prelude setup, we are moving the custom code back
into Nix.Utils, so the downstream projects do not have trouble importing Prelude
module, but just use regular Nix.Utils module.
